Initial File Move (Read Description)
We orginally misintrepeted where to host the girhub, and as such had done most of the work on that github. That github had about 20 commits and two branches. This first push is just moving all the approprate files to the proper postion. The later pushes will be updating the JS and HTML to deal with the new file strucute.